chore: refactor `<PreventNavigationOnUnsavedChanges>` by extracting a `usePreventNavigationOnUnsavedChanges` hook
